# [Benchmark] Allow oversample request in benchmark dataset