Enable oversample requests in HuggingFaceDataset, VisionArenaDataset and ShareGPTDataset.

This is mostly needed for small datasets like visionarena. 

**This branch**

| Dataset | Backend | Successful requests | Benchmark duration (s) | Total input tokens |
|---------|---------|---------------------|--------------------------|--------------------|
| hf-vision-arena | openai-chat | 10000      | 444.00     | 627745     |
| hf | openai-chat | 10000      | 490.06     | 114223     |
| sharegpt | vllm | 10000      | 262.65     | 2205227    |

**Main branch**

| Dataset | Backend | Successful requests | Benchmark duration (s) | Total input tokens |
|---------|---------|---------------------|--------------------------|--------------------|
| hf-vision-arena | openai-chat | 500        | 36.29      | 33418      |
| hf | openai-chat | 10000      | 413.91     | 114223     |
| sharegpt | vllm | 10000      | 297.65     | 2205227    |

```
vllm serve Qwen/Qwen2-VL-7B-Instruct     --swap-space 16     --disable-log-requests
```

```
python3 vllm/benchmarks/benchmark_serving.py --model Qwen/Qwen2-VL-7B-Instruct --backend openai-chat --endpoint /v1/chat/completions --dataset-name hf --dataset-path lmarena-ai/vision-arena-bench-v0.1 --hf-split train --num-prompts 10000 --percentile-metrics ttft,tpot,e2el

python3 vllm/benchmarks/benchmark_serving.py --model Qwen/Qwen2-VL-7B-Instruct --backend openai-chat --endpoint /v1/chat/completions --dataset-name hf --dataset-path lmms-lab/LLaVA-OneVision-Data --hf-split train --hf-subset "chart2text(cauldron)" --num-prompts 10000 --percentile-metrics ttft,tpot,e2el

python3 vllm/benchmarks/benchmark_serving.py --backend vllm --model Qwen/Qwen2-VL-7B-Instruct --dataset-name sharegpt --dataset-path /home/jovyan/data/vllm_benchmark_datasets/ShareGPT_V3_unfiltered_cleaned_split.json --num-prompts 10000
```
<!--- pyml disable-next-line no-emphasis-as-heading -->